+++ src/main/java/akkodis/tryout/prices/domain/model/Prices.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -29,7 +28,7 @@ public class Prices {
 
     private Integer priority;
 
-    private BigDecimal price;
+    private Float price;
 
     private String curr;
