+++ src/main/java/akkodis/tryout/prices/infrastructure/repositories/entities/PricesEntity.java
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -43,7 +42,7 @@ public class PricesEntity {
     private Integer priority;
 
     @Column(name = "price")
-    private BigDecimal price;
+    private Float price;
 
     @Column(name = "curr")
     private String curr;
